backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
openpyxl>=3.1.2
python-multipart>=0.0.9
fpdf2>=2.7.9
uharfbuzz>=0.39.0
fonttools>=4.51.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
openpyxl>=3.1.2
python-multipart>=0.0.9
fpdf2>=2.7.9
uharfbuzz>=0.39.0
fonttools>=4.51.0
pytest>=8.2.0
xlrd>=2.0.1
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5.0.0
slowapi
httpx
